requirements/local.txt
-r ./base.txt

# Debugging
ipdb==0.11

# Tools
django-extensions==2.1.4

# Testing
mypy==0.650
pytest==6.2.4
pytest-sugar==0.9.4
pytest-django==4.3.0
factory-boy==2.11.1

# Code quality
flake8==3.6.0
